Fix issues introduced by mod tap
mod tap only supports 8-bit keycodes, which means having mod tap on both
'0' and ')' does not work, since RPAREN is the shifted version of 0. On
both keys, the mod-tapped mod was RSFT (not sure if it matters if they
were the same).

Initially I tried to circumvent this with tap dance (there's an example
that simulates mod tap), but it's hard to precisely emulate mod tap
behavior with it (e.g. repeatedly tapping the key will behave
differently)

Interestingly, this issue does not occur with '-' and '_' (here,
underscore is the shifted version of minus). Did not look into why.
Maybe it was because the modifiers were different?

My solution was to dodge the bullet and modify the keymap, so RPAREN is
no longer on mod-tap position.